requirements.txt
tensorflow-gpu>=1.13.1
keras>=2.2.4
opencv-python>=3.4.0.12
pyaudio>=0.2.11
Pillow>=5.3.0
librosa>=0.6.3
sidekit>=1.3.2
lxml
matplotlib
time
pydub
wave
numpy
scipy
pandas